+++ bindings/python/include/aligator/python/fwd.hpp
@@ -45,6 +45,7 @@ void exposeCallbacks();
 /// Expose autodiff helpers
 void exposeAutodiff();
 void exposeUtils();
+void exposeFilter();
 
 #ifdef ALIGATOR_WITH_PINOCCHIO
 /// Expose features using the Pinocchio rigid dynamics library