+++ bindings/python/src/expose-solver-prox.cpp
@@ -133,6 +133,8 @@ void exposeProxDDP() {
       .def_readwrite("max_al_iters", &SolverType::max_al_iters,
                      "Maximum number of AL iterations.")
       .def_readwrite("ls_mode", &SolverType::ls_mode, "Linesearch mode.")
+      .def_readwrite("sa_strategy", &SolverType::sa_strategy,
+                     "StepAcceptance strategy.")
       .def_readwrite("rollout_type", &SolverType::rollout_type_,
                      "Rollout type.")
       .def_readwrite("dual_weight", &SolverType::dual_weight,
@@ -144,6 +146,8 @@ void exposeProxDDP() {
       .def("updateLQSubproblem", &SolverType::updateLQSubproblem, "self"_a)
       .def("computeCriterion", &SolverType::computeCriterion, "self"_a,
            "Compute problem stationarity.")
+      .def_readwrite("filter", &SolverType::filter_,
+                     "Pair filter used to accept a step.")
       .def("computeInfeasibilities", &SolverType::computeInfeasibilities,
            ("self"_a, "problem"), "Compute problem infeasibilities.")
       .def(SolverVisitor<SolverType>())