+++ bindings/python/src/module.cpp
@@ -38,6 +38,12 @@ static void exposeEnums() {
       .value("HESSIAN_EXACT", HessianApprox::EXACT)
       .value("HESSIAN_GAUSS_NEWTON", HessianApprox::GAUSS_NEWTON)
       .export_values();
+
+  bp::enum_<StepAcceptanceStrategy>("StepAcceptanceStrategy",
+                                    "Step acceptance strategy.")
+      .value("SA_LINESEARCH", StepAcceptanceStrategy::LINESEARCH)
+      .value("SA_FILTER", StepAcceptanceStrategy::FILTER)
+      .export_values();
 }
 
 static void exposeContainers() {
@@ -75,6 +81,7 @@ BOOST_PYTHON_MODULE(MODULE_NAME) {
   exposeConstraint();
   exposeStage();
   exposeProblem();
+  exposeFilter();
   {
     bp::scope dynamics = get_namespace("dynamics");
     exposeODEs();