+++ examples/talos-arm.cpp
@@ -2,7 +2,6 @@
 
 #include "aligator/solvers/proxddp/solver-proxddp.hpp"
 #include "aligator/compat/crocoddyl/problem-wrap.hpp"
-#include <proxsuite-nlp/tests/cnpy.hpp>
 #include <fmt/ostream.h>
 
 using Eigen::MatrixXd;