+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Document and simplify the LQR example
 - Finish the se2-car.cpp example #110
 - Add template instantiation for IntegratorAbstract, ExplicitIntegratorAbstract and IntegratorRK2 #114
+- Fix for eigenpy 3.3
+- Don't output numpy matrices in `example-talos-arm`
 
 ### Fixed
 
@@ -24,6 +26,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] - 2023-12-22
 
 ### Added
+
+- Add a pair filter strategy as an alternative to linesearch methods
+- Add a python example of a locomotion OCP with the robot Talos
 
 * This is the first release of `aligator`. This library is a joint effort between INRIA and LAAS-CNRS, and will be maintained and expanded in the future. Please provide constructive feedback and contribute!